requirements.txt
Authlib==1.6.9
beautifulsoup4==4.14.3
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.5
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
filelock==3.25.2
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
greenlet==3.3.2
gunicorn==25.1.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
marshmallow==4.2.2
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
psycopg2-binary==2.9.10
pytest==9.0.2
pytest-cov==7.1.0
pytest-flask==1.3.0
python-dotenv==1.0.1
python-json-logger==4.0.0
redis==7.3.0
requests==2.32.5
ruff==0.15.7
sentry-sdk==2.55.0
soupsieve==2.8.3
SQLAlchemy==2.0.36
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
whitenoise==6.8.2
wrapt==2.1.2
WTForms==3.2.1
